backend/data/requirements.txt
Flask-SocketIO==3.1.1
eventlet==0.24.1
websocket-client==0.52.0
pyproj==1.9.5.1
requests
kombu
aiohttp
cchardet
aiodns
